+++ src/main/java/simba/ui/EmptyException.java
@@ -1,0 +1,10 @@
+package simba.ui;
+
+/**
+ * Thrown when a task description is empty.
+ */
+public class EmptyException extends Exception {
+    public EmptyException(String msg) {
+        super(msg);
+    }
+}
