+++ src/main/java/simba/ui/NonsenseException.java
@@ -1,0 +1,7 @@
+package simba.ui;
+
+/**
+ * Thrown when an invalid command is entered.
+ */
+public class NonsenseException extends Exception {
+}
