+++ src/test/java/simba/ui/SimbaTest.java
@@ -1,22 +1,20 @@
-package simba.ui;  //same package as the class being tested
+package simba.ui;
 
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import java.util.ArrayList;
-
 public class SimbaTest {
-    Parser parser1 = new Parser("delete 1");
-    Parser parser2 = new Parser("delete 2");
 
     @Test
-    public void dummyTest(){
+    public void parserTest1(){
+        Parser parser1 = new Parser("delete 1");
         assertEquals(1, parser1.idxToDelete());
     }
 
     @Test
-    public void anotherDummyTest(){
+    public void parserTest2(){
+        Parser parser2 = new Parser("delete 2");
         assertEquals(2, parser2.idxToDelete());
     }
 }
